Add parser for select choices
